requirements.txt
flask
flask-restful
pyyaml